While restoring, rrd_restore uses between 30 and 40 times as much memory as the size of the rrd file to be restored. This is due to the use of libxml2 for parsing the the whole file prior to turning it into rrd structs.
This patch switches rrd_restore to using the xmlTextReader interface for
incremental consumtion of xml files. This reduces the memory requirements to
about twice the size of the rrd file which should be quite manageable for
most cases.

The parsing is about 25% less efficient than with the original libxml2
version of rrd_restore. This could be in part due to the more liberal use of
malloced memory. Optimizing this is for another day though.

--tobi