rrd_tune HW Update
Allow to set the smoothing window (alpha, beta, gamma) to zero.

I'm using rrdtool 1.3.8 with abberant behavior detection from the Brutlag
method. I need to use a week seasonality and no smoothing for the SEASONAL
and DEVSEASONAL. So I use rrdtune in order to set the smoothing window to
0.0. This calls the set_hwarg(). But this method is also used to tune the
alpha, beta and gamma that can't be set to 0.0 So when I tune my smoothing
window size to 0.0, it returns -1 due to the if(window<=0.0) then do
nothing. I join here a small correction with a differnent method used to set
only the smoothing window size and allowing it to be zero. 

-- sylvain.luiset  gmail.com