rrd_client: Do not rewrite path names when accessing remote daemons.
When talking to a local daemon (thru a UNIX socket), relative path names are
resolved to absolute path names to allow for transparent integration into
existing solutions (as requested by Tobi).

However, when talking to a remote daemon, absolute path names are not allowed,
since path name translation is done by the server (relative to the base
directory). -- Sebastian Harl