src/rrd_client.c: Fix a potential segmentation fault in "get_path".
The "get_path" function is called at least from "rrdc_update" without
checking for a connection first. In that case the "sd_path" pointer may be
NULL, so dereferencing it without checking for NULL is not good behavior.

This patch checks all arguments of the "get_path" function and returns an
error if any pointer is NULL. -- Florian Forster